Moving cornersToString() into a toString() overloading to be consistent.  Fixed a few compiler warnings.
git-svn-id: https://wush.net/svn/mdaus-oss/trunk@370 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/six/source/NITFReadControl.cpp
@@ -251,7 +251,7 @@ void NITFReadControl::load(nitf::IOInterface& interface)
 
         // This function should throw if the data does not exist
         std::pair<int, int> imageAndSegment = getIndices(subheader);
-        if (imageAndSegment.first >= mInfos.size())
+        if (static_cast<size_t>(imageAndSegment.first) >= mInfos.size())
         {
             throw except::Exception(Ctxt(
                     "Image " + str::toString(imageAndSegment.first) +
